+++ Configula/Config/PluginConfig.cs
@@ -10,8 +10,16 @@ public static class PluginConfig {
   public static ConfigEntry<int> WindowWidth { get; private set; }
   public static ConfigEntry<int> WindowHeightOffset { get; private set; }
 
+  public static ConfigEntry<bool> UseAlternateDrawTooltip { get; private set; }
+  public static ConfigEntry<int> AlternateTooltipWidth { get; private set; }
+
   public static void BindConfig(ConfigFile config) {
-    IsModEnabled = config.BindInOrder("_Global", "isModEnabled", true, "Globally enable or disable this mod.");
+    IsModEnabled =
+        config.BindInOrder(
+            "_Global",
+            "isModEnabled",
+            true,
+            "Globally enable or disable this mod (restart required).");
 
     WindowWidth =
         config.BindInOrder(
@@ -28,5 +36,20 @@ public static class PluginConfig {
             100,
             "Height (offset) of the ConfigurationManager window.",
             new AcceptableValueRange<int>(0, 540));
+
+    UseAlternateDrawTooltip =
+        config.BindInOrder(
+            "Tooltip",
+            "useAlternateDrawTooltip",
+            true,
+            "If set, will use alternative DrawTooltip() for customization and Unity-v6 compatibility.");
+
+    AlternateTooltipWidth =
+        config.BindInOrder(
+            "Tooltip",
+            "alternateTooltipWidth",
+            400,
+            "Width of the alternate tooltip (clamped to window width).",
+            new AcceptableValueRange<int>(100, 1000));
   }
 }